+++ src/main/java/com/nixinova/morecmds/commands/Home.java
@@ -20,9 +20,11 @@ import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.mojang.brigadier.exceptions.SimpleCommandExceptionType;
 import net.fabricmc.fabric.api.command.v1.CommandRegistrationCallback;
+import net.minecraft.client.MinecraftClient;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.TranslatableText;
+import net.minecraft.util.WorldSavePath;
 import net.minecraft.util.math.BlockPos;
 
 import com.nixinova.morecmds.Main;
@@ -32,14 +34,15 @@ import com.nixinova.morecmds.Permission;
 public class Home {
 
 	static final File homesFile = Main.CONFIG_DIR.resolve("homes.txt").toFile();
-	static Map<String, float[]> homes = new HashMap<>();
+
+	static Map<String, Map<String, float[]>> homesData = new HashMap<>();
 
 	private enum ArgType {
 		SET, REMOVE, GET, GO, LIST
 	}
 
 	public void register() {
-		Home.setupConfig();
+		setupConfig();
 		CommandRegistrationCallback.EVENT.register((dispatcher, dedicated) -> {
 			dispatcher.register(
 				literal("home").then(
@@ -82,6 +85,11 @@ public class Home {
 			return -1;
 		}
 
+		String world = getWorld();
+		if (!homesData.containsKey(world)) {
+			homesData.put(world, new HashMap<>());
+		}
+
 		switch (type) {
 			case SET -> setHome(context);
 			case REMOVE -> removeHome(context);
@@ -98,23 +106,23 @@ public class Home {
 		BlockPos pos = getBlockPos(context, "pos");
 		float[] coords = {pos.getX(), pos.getY(), pos.getZ()};
 
-		homes.put(name, coords);
-		writeConfig(name);
+		homesData.get(getWorld()).put(name, coords);
+		writeConfig();
 
 		ServerPlayerEntity player = context.getSource().getPlayer();
-		Messages.generic("success.home.home.set", player, name, coords[0], coords[1], coords[2]);
+		Messages.generic("success.home.set", player, name, coords[0], coords[1], coords[2]);
 	}
 
 	private void removeHome(CommandContext<ServerCommandSource> context) throws CommandSyntaxException {
 		String name = getString(context, "name");
-		float[] coords = homes.get(name);
+		float[] coords = homesData.get(getWorld()).get(name);
 		if (coords == null) {
 			TranslatableText invalid = new TranslatableText("command.error.home.notFound", name);
 			throw new SimpleCommandExceptionType(invalid).create();
 		}
 
-		homes.remove(name);
-		writeConfig(name);
+		homesData.get(getWorld()).remove(name);
+		writeConfig();
 
 		ServerPlayerEntity player = context.getSource().getPlayer();
 		Messages.generic("success.home.remove", player, name);
@@ -122,7 +130,7 @@ public class Home {
 
 	private void getHome(CommandContext<ServerCommandSource> context) throws CommandSyntaxException {
 		String name = getString(context, "name");
-		float[] coords = homes.get(name);
+		float[] coords = homesData.get(getWorld()).get(name);
 		if (coords == null) {
 			TranslatableText invalid = new TranslatableText("command.error.home.notFound", name);
 			throw new SimpleCommandExceptionType(invalid).create();
@@ -133,7 +141,7 @@ public class Home {
 
 	private void goHome(CommandContext<ServerCommandSource> context) throws CommandSyntaxException {
 		String name = getString(context, "name");
-		float[] coords = homes.get(name);
+		float[] coords = homesData.get(getWorld()).get(name);
 		if (coords == null) {
 			TranslatableText invalid = new TranslatableText("command.error.home.notFound", name);
 			throw new SimpleCommandExceptionType(invalid).create();
@@ -146,7 +154,7 @@ public class Home {
 	private void listHomes(CommandContext<ServerCommandSource> context) throws CommandSyntaxException {
 		ServerPlayerEntity player = context.getSource().getPlayer();
 		int total = 0;
-		for (Map.Entry<String, float[]> entry : homes.entrySet()) {
+		for (Map.Entry<String, float[]> entry : homesData.get(getWorld()).entrySet()) {
 			String name = entry.getKey();
 			float[] coords = entry.getValue();
 			Messages.generic("info.home.list.entry", player, name, coords[0], coords[1], coords[2]);
@@ -155,21 +163,33 @@ public class Home {
 		Messages.generic("info.home.list.total", player, total);
 	}
 
+	private static String getWorld() {
+		MinecraftClient client = MinecraftClient.getInstance();
+		String world = client.getServer().getSavePath(WorldSavePath.ROOT).getParent().getFileName().toString();
+		return world;
+	}
+
 	public static void setupConfig() {
 		Main.log("Setting up config file 'homes.txt'");
 		try {
 			Scanner homesReader = new Scanner(homesFile);
 			while (homesReader.hasNextLine()) {
 				String line = homesReader.nextLine();
-				if (!line.matches(".+:\\d+\\.\\d+/\\d+\\.\\d+/\\d+\\.\\d+")) continue;
+				if (!line.matches(".+?:.+?:\\d+\\.\\d+/\\d+\\.\\d+/\\d+\\.\\d+")) {
+					continue;
+				}
 				String[] data = line.split(":");
-				String name = data[0];
+				String world = data[0];
+				if (world != MinecraftClient.getInstance().world.toString()) {
+					return;
+				}
+				String name = data[1];
 				String[] coordsData = data[1].split("/");
 				float[] coords = {0, 0, 0};
 				for (int i = 0; i < coordsData.length; i++) {
 					coords[i] = Float.parseFloat(coordsData[i]);
 				}
-				homes.put(name, coords);
+				homesData.get(getWorld()).put(name, coords);
 			}
 			homesReader.close();
 			Main.log("Successfully loaded from config file 'homes.txt'");
@@ -180,19 +200,22 @@ public class Home {
 		}
 	}
 
-	private void writeConfig(String name) {
+	private void writeConfig() {
 		try {
 			if (homesFile.createNewFile()) {
 				Main.log("Created config file 'homes.txt'");
 			}
 			FileWriter homesFileBuffer = new FileWriter(homesFile);
-			for (Map.Entry<String, float[]> entry : homes.entrySet()) {
-				String homeName = entry.getKey();
-				float[] homeCoords = entry.getValue();
-				homesFileBuffer.write(String.format("%s:%f/%f/%f\n", homeName, homeCoords[0], homeCoords[1], homeCoords[2]));
+			for (String world : homesData.keySet()) {
+				for (Map.Entry<String, float[]> entry : homesData.get(world).entrySet()) {
+					String homeName = entry.getKey();
+					float[] homeCoords = entry.getValue();
+					String line = String.format("%s:%s:%f/%f/%f", world, homeName, homeCoords[0], homeCoords[1], homeCoords[2]);
+					homesFileBuffer.write(line + "\n");
+				}
 			}
 			homesFileBuffer.close();
-			Main.log(String.format("Edited data for home '%s' in config file 'homes.txt'", name));
+			Main.log("Edited config file 'homes.txt'");
 		}
 		catch (IOException err) {
 			Main.log("Error while creating config file 'homes.txt'");